Optimize the deleteJob function

+++ controllers/jobController.js
@@ -62,10 +62,16 @@ exports.getById = async (req , res) => {
 // Delete a Job
 exports.deleteJob = async (req, res) => {
     try{
-        const job = await Job.findByIdAndDelete(req.params.id);
+        const job = await Job.findById(req.params.id);
         if(!job){
             return res.status(404).json({message: "Job not Found!"});
         }
+
+        if(job.postedBy.toString() !== req.user.id){
+            return res.status(401).json({message: "Unauthorized Delete!"});
+        }
+
+        await Job.findByIdAndDelete(req.params.id);
         res.status(200).json({message: "Job Deleted Successfully"});
     }
     catch(err){
